PNG: Fix casting (LGTM)

+++ src/sail-codecs/png/png.c
@@ -418,7 +418,7 @@ SAIL_EXPORT sail_status_t sail_codec_read_frame_v5_png(void *state, struct sail_
                 if (png_state->next_frame_dispose_op == PNG_DISPOSE_OP_BACKGROUND) {
                     memset(png_state->prev[row] + png_state->next_frame_x_offset * png_state->bytes_per_pixel,
                             0,
-                            png_state->next_frame_width * png_state->bytes_per_pixel);
+                            (size_t)png_state->next_frame_width * png_state->bytes_per_pixel);
                 } else if (png_state->next_frame_dispose_op == PNG_DISPOSE_OP_NONE) {
                     memcpy(png_state->prev[row] + png_state->next_frame_x_offset * png_state->bytes_per_pixel,
                             scanline,
